move DifferingRunsTable to its own file
