+++ frontends/ol-components/src/components/LearningResourceExpanded/InfoSectionV2.test.tsx
@@ -5,7 +5,6 @@ import InfoSectionV2 from "./InfoSectionV2"
 import { ThemeProvider } from "../ThemeProvider/ThemeProvider"
 import { formatRunDate } from "ol-utilities"
 import invariant from "tiny-invariant"
-import { DeliveryEnumDescriptions } from "api"
 
 // This is a pipe followed by a zero-width space
 const SEPARATOR = "|​"
@@ -155,38 +154,5 @@ describe("Learning resource info section start date", () => {
     within(section).getByText((_content, node) => {
       return node?.textContent === expectedDateText || false
     })
-  })
-})
-describe("Differing runs comparison table", () => {
-  test("Does not appear if data is the same", () => {
-    const course = courses.multipleRuns.sameData
-    render(<InfoSectionV2 resource={course} />, {
-      wrapper: ThemeProvider,
-    })
-    expect(screen.queryByTestId("differing-runs-table")).toBeNull()
-  })
-
-  test("Appears if data is different", () => {
-    const course = courses.multipleRuns.differentData
-    render(<InfoSectionV2 resource={course} />, {
-      wrapper: ThemeProvider,
-    })
-    const differingRunsTable = screen.getByTestId("differing-runs-table")
-    expect(differingRunsTable).toBeInTheDocument()
-    const onlineLabels = within(differingRunsTable).getAllByText(
-      DeliveryEnumDescriptions.online,
-    )
-    const inPersonLabels = within(differingRunsTable).getAllByText(
-      DeliveryEnumDescriptions.in_person,
-    )
-    const onlinePriceLabels =
-      within(differingRunsTable).getAllByText("$0, $100")
-    const inPersonPriceLabels = within(differingRunsTable).getAllByText("$150")
-    const earthLocationLabels = within(differingRunsTable).getAllByText("Earth")
-    expect(onlineLabels).toHaveLength(2)
-    expect(inPersonLabels).toHaveLength(2)
-    expect(onlinePriceLabels).toHaveLength(2)
-    expect(inPersonPriceLabels).toHaveLength(2)
-    expect(earthLocationLabels).toHaveLength(2)
   })
 })
